requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
cameratransform==1.2
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.11.0
defer==1.0.6
distro==1.9.0
distro-info===0.18ubuntu0.18.04.1
exif==1.2.2
feedparser==5.2.1
geographiclib==1.52
geopy==2.2.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-resources==5.4.0
Jetson.GPIO==2.0.4
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.3.4
numpy==1.19.4
oauth==1.0.1
oauthlib==2.0.6
packaging==21.3
PAM==0.4.2
pandas==1.1.5
Pillow==8.4.0
plum-py==0.3.1
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.1.4
pyRFC3339==1.0
pyserial==3.5
python-apt==1.6.4
python-dateutil==2.9.0.post0
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
scikit-build==0.16.7
scipy==1.5.4
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sparkfun-ublox-gps==1.1.5
spidev==3.6
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tqdm==4.64.1
typing_extensions==4.1.1
ubuntu-drivers-common==0.0.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
wadllib==1.3.2
webencodings==0.5
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
